Pull request 582: Exclude some clang-tidy checks and fix all remaining tidy errors
Squashed commit of the following:

commit 7630effb7815b14aeaf833ba9baf649a4a197b3c
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 17:27:19 2026 +0300

    Add http code to http status enum value name

commit 4015290c7c03d8bdd4e88b1e2f99bd2ef94a8bfd
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 15:22:50 2026 +0300

    Exclude concurrency-mt-unsafe check due to inability to turn it off for strerror

commit 8ee2a2872bb2751eeebd6b7e331ad09d35d16146
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 14:04:38 2026 +0300

    Fix (nolint) bugprone-unchecked-optional-access

commit 68e161c5487019268f63b54a0edabdbe87841695
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:13:06 2026 +0300

    Fix bugprone-optional-value-conversion

commit 0dd29f93aa68aa14900db7ce878de63b30789f9a
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:11:25 2026 +0300

    Fix bugprone-unused-return-value

commit 58e749a8fe2261c40481b8080ab6ed06fd82741f
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:06:13 2026 +0300

    Fix cppcoreguidelines-misleading-capture-default-by-value

commit 9355974fd3d550d6b854187c04b65039eaaf46e0
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:05:12 2026 +0300

    Fix cppcoreguidelines-no-malloc

commit 15ffd925a54af42051c6075d92b6021c1eeec9e0
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:04:21 2026 +0300

    Fix readability-container-size-empty

commit 28c1e6102f550be625356a8e657600a9a226e6b6
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 13:03:44 2026 +0300

    Fix bugprone-use-after-move

commit 6f20ad04c5412da9b88acd9ea03f140abaaba677
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:59:53 2026 +0300

    Fix hicpp-use-emplace

commit 61a6b91c5dce91acfcf13f7836d20089d76a611b
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:51:10 2026 +0300

    Fix performance-avoid-endl

commit 5c7c8eaf0016d988999be4994374c3076444b0ce
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:50:22 2026 +0300

    Fix readability-avoid-return-with-void-value

commit 75ef9f28624c1f23bc8a80edc4953342bbfa9667
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:49:08 2026 +0300

    Fix readability-container-contains

commit dc09e3a1e56874ff4c6f803addf09c975e925dfe
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:47:57 2026 +0300

    Fix bugprone-narrowing-conversions

commit 9c465c0650b7e84eb8805085dadadb3e1cba3579
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:37:47 2026 +0300

    Fix hicpp-move-const-arg

commit 818abcf1596e21e14d7031de04d8395d4f05950f
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:34:00 2026 +0300

    Fix readability-container-data-pointer

commit fb7ed6ce210cbb440c58da3bcb0a41df8a19ad01
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:29:35 2026 +0300

    Fix cppcoreguidelines-avoid-magic-numbers

commit 40f58aa179f5d0e1c8f66ac92eaedade887c4c04
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 12:10:34 2026 +0300

    Sync cmake flags between setup_cmake and compile_commands rules in Makefile

commit cc0dffe1325854dd330529229b556ade87466a71
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Fri Jan 16 10:14:55 2026 +0300

    Apply review suggestions

commit e2b951a82c651c2625f5e8791cde2b978d227d58
Author: Zhavoronkov Aleksei <a.zhavoronkov@adguard.com>
Date:   Thu Jan 15 17:03:46 2026 +0300

    Fix clang-format

... and 31 more commits

+++ tcpip/src/tcp_conn_manager.cpp
@@ -232,7 +232,7 @@ int tcp_cm_send_data(TcpConnDescriptor *connection, const uint8_t *data, size_t 
 
     tcp_refresh_connection_timeout(connection);
 
-    return bytes_to_send;
+    return static_cast<int>(bytes_to_send);
 }
 
 void tcp_cm_data_sent_notify(TcpConnDescriptor *connection, size_t length) {
@@ -268,7 +268,7 @@ void tcp_cm_close_descriptor(TcpipCtx *ctx, uint64_t id, bool graceful) {
     log_conn(connection, trace, "Connection closed {}, {} active connections left", (void *) connection,
             kh_size(ctx->tcp.connections.by_id));
 
-    free(connection);
+    free(connection); // NOLINT(cppcoreguidelines-no-malloc,hicpp-no-malloc)
 }
 
 int tcp_cm_receive(TcpConnDescriptor *connection, size_t iovlen, const evbuffer_iovec *iov) {
@@ -284,6 +284,7 @@ int tcp_cm_receive(TcpConnDescriptor *connection, size_t iovlen, const evbuffer_
 TcpConnDescriptor *tcp_cm_create_descriptor(TcpipCtx *ctx, struct pbuf *buffer, const ip_addr_t *src_addr,
         u16_t src_port, const ip_addr_t *dst_addr, u16_t dst_port) {
     static_assert(std::is_trivial_v<TcpConnDescriptor>);
+    // NOLINTNEXTLINE(cppcoreguidelines-no-malloc,hicpp-no-malloc)
     auto *connection = (TcpConnDescriptor *) calloc(1, sizeof(TcpConnDescriptor));
     if (nullptr == connection) {
         return nullptr;
